Fix Game Over Display not showing
Issue: When game ends in a draw, player objects need to be fetched to
display their name and avatar. Due to `_endGameScreen` being an `async`
function, `endGame` in the `gameController` object does not wait for
the Game Over Display to be shown before removing the players.

`endGame` is now called on fulfillment of the `_endGameScreen` Promise